// src/app/components/charts/chart-execute-bot/shared-chart-execute-bot/interfaces/shared-chart-execute-bot.interface.ts
import {
    ApexAxisChartSeries,
    ApexChart,
    ApexDataLabels,
    ApexYAxis,
    ApexLegend,
    ApexXAxis,
    ApexTooltip,
    ApexTheme,
    ApexGrid,
  } from "ng-apexcharts";

export type ExecuteBotChartOptions = {
    series: ApexAxisChartSeries;
    chart: ApexChart;
    xaxis: ApexXAxis;
    yaxis: ApexYAxis;
    stroke: any;
    theme: ApexTheme;
    tooltip: ApexTooltip;
    dataLabels: ApexDataLabels;
    legend: ApexLegend;
    colors: string[];
    markers: any;
    grid: ApexGrid;
  };